Moved JPB breadcrumb code from router into JobBuilderRoot file
